feat: acc=0.82.

+++ Classification.MNIST/utils.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 
-def arrayToBits(array, n=8, dtype=np.int32):
+def arrayToBits(array, n=8, dtype=np.uint8):
     H, W = array.shape[:2]
     array = np.array(array, dtype=dtype)
     bits = np.zeros((H, W, n), dtype=dtype)
@@ -10,7 +10,7 @@ def arrayToBits(array, n=8, dtype=np.int32):
     return bits
 
 
-def bitsToarray(bits, dtype=np.int32):
+def bitsToarray(bits, dtype=np.uint8):
     H, W, n = bits.shape
     bits = np.array(bits, dtype=dtype)
     array = np.zeros((H, W), dtype=dtype)
